Add files via upload
As we know, Chicago is of the largest cities in the world. It is known as windy city. Not only It is
famous for its food and skyline but also it is one of the cities in USA with highest crime rates.
Crime is a significant social problem in the nation, affecting public safety, child development,
and adult socioeconomic status. Chicago has always been a holiday destination for international
tourists and top pick for international students to pursue their studies. Safety is the primary
concern while staying in a new city, miles away from home. The motivation behind taking up this
topic is to use predictive techniques to assess the safety of a place and feel safe in our
neighborhood.